+++ src/compas_cadwork/scene/instructionobject.py
@@ -1,4 +1,6 @@
 import cadwork
+
+from compas.geometry import Frame
 from compas_monosashi.sequencer import LinearDimension
 from compas_monosashi.sequencer import Model3d
 from compas_monosashi.sequencer import Text3d
@@ -6,11 +8,11 @@ from dimension_controller import create_dimension
 from element_controller import apply_transformation_coordinate
 from element_controller import create_text_object_with_options
 from element_controller import get_bounding_box_vertices_local
-from element_controller import move_element
 from file_controller import import_element_light
 
 from compas_cadwork.conversions import point_to_cadwork
 from compas_cadwork.conversions import vector_to_cadwork
+from compas_cadwork.conversions import point_to_compas
 from compas_cadwork.scene import CadworkSceneObject
 
 
@@ -30,47 +32,41 @@ class Text3dSceneObject(CadworkSceneObject):
         self.text_instruction = text_instruction
 
     @staticmethod
-    def _generate_translation_vectors(element_id: int):
-        """Generates translation vectors from a bounding box that shift a text
-        or a box from the bottom left point of the object to the center point
-        of the object.
+    def _generate_translation_vectors(element_id: int, inst_frame: Frame):
+        # calculates the translation vectors needed properly center the text
+        width, height = Text3dSceneObject._calculate_text_size(element_id)
+        shift_x = inst_frame.xaxis.scaled(-0.5 * width)
+        shift_y = inst_frame.yaxis.scaled(-0.5 * height)
 
-        Parameters
-        ----------
-        element_ids : int
-            Cadwork element id of the text object.
+        # shift text 5 mm in z direction to ensure it does not get hidden inside the element
+        shift_z = inst_frame.normal.scaled(5.0)
+        return shift_x + shift_y + shift_z
 
-        Return
-        -------
-        tuple(cadwork.point_3d, cadwork.point_3d, cadwork.point_3d)
-            Translation vectors in x, y and z direction.
-        """
+    @staticmethod
+    def _calculate_text_size(element_id: int):
+        # use the bounding box to determine the size of the text
+        #  0 -------- 2
+        #  ^          |
+        #  h          |
+        #  |          |
+        #  1 --------w> 3
         bb = get_bounding_box_vertices_local(element_id, [element_id])
+        p0 = point_to_compas(bb[0])
+        p1 = point_to_compas(bb[1])
+        p3 = point_to_compas(bb[3])
+        d1 = p1.distance_to_point(p3)
+        d2 = p0.distance_to_point(p1)
 
-        # https://github.com/inconai/innosuisse_issue_collection/issues/137
-        start_vec_x = bb[0]
-        end_vec_x = bb[1]
-        start_vec_z = bb[0]
-        end_vec_z = bb[2]
-
-        vx = start_vec_x - end_vec_x
-        dx = start_vec_x.distance(end_vec_x) / 2.0
-
-        vx = vx.normalized()
-        vx = vx * dx
-
-        vz = start_vec_z - end_vec_z
-        dz = start_vec_z.distance(end_vec_z) / 2.0
-
-        vz = vz.normalized()
-        vz = vz * dz
-
-        # shift text in vy so it doesnt intersect with geometry
-        vy = vz.cross(vx)
-        vy = vy.normalized()
-        vy = vy * 5
-
-        return vx, vy, vz
+        # https://github.com/inconai/innosuisse_issue_collection/issues/259
+        # this is a hack designed to get over the inconsistency of the bounding box's orientation
+        # the texts are longer then they are tall, so determine which is which depending on the ratio.
+        if d1 > d2:
+            width = d1
+            height = d2
+        else:
+            width = d2
+            height = d1
+        return width, height
 
     def draw(self, *args, **kwargs):
         """Adds a text element with the text included in the provided text instruction.
@@ -94,11 +90,13 @@ class Text3dSceneObject(CadworkSceneObject):
         element_id = create_text_object_with_options(
             point_to_cadwork(loc.point), vector_to_cadwork(loc.xaxis), vector_to_cadwork(loc.yaxis), text_options
         )
-        if self.text_instruction.centered:
-            vx, vy, vz = self._generate_translation_vectors(element_id)
-            move_element([element_id], vx + vy + vz)
 
         element = self.add_element(element_id)
+
+        if self.text_instruction.centered:
+            translation = self._generate_translation_vectors(element_id, self.text_instruction.location)
+            element.translate(translation)
+
         element.set_is_instruction(True, self.text_instruction.id)
         return [element_id]