+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added method `translate` to `Element`.
+
 ### Changed
+
+* Fixed centered text instructions are sometimes shifted the worng direction.
 
 ### Removed